Do not delete package folders if they are symlinks

+++ src/Mover.php
@@ -139,6 +139,9 @@ class Mover
     {
         foreach ($this->movedPackages as $movedPackage) {
             $packageDir = '/vendor/' . $movedPackage;
+            if (is_link($packageDir)) {
+            	continue;
+            }
             $this->filesystem->deleteDir($packageDir);
         }
     }
